# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: AR Demo and candidate API
  version: 0.2.0
  description: 'Demo: RuoYi + PostgreSQL, fixed-role administrators and scenes. Docker
    Compose and local ArtifactStorage infrastructure are available without artifact
    HTTP endpoints. Historical cloud/day-night resource and publication contracts
    below await a real deliverable and client loading agreement; they are not implemented.'
servers:
- url: https://api.example.invalid
  description: placeholder
paths:
...
  /system/user/profile:
    get:
      operationId: profile
      summary: 本人资料
      description: 本人资料。沿用若依 code/msg 响应；业务错误可能使用 HTTP 200。
      security:
      - bearerAuth: []
      parameters: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
        '400':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
        '401':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
        '403':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
        '404':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
        '409':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
        '413':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
        '429':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
        '503':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
      x-implementation: implemented
      tags:
      - Demo 已实现
    put:
      operationId: updateProfile
      summary: 修改本人资料
      description: 修改本人资料。沿用若依 code/msg 响应；业务错误可能使用 HTTP 200。
      security:
      - bearerAuth: []
      parameters: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
        '400':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
        '401':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
        '403':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
        '404':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
        '409':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
        '413':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
        '429':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
        '503':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
      x-implementation: implemented
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileUpdate'
      tags:
      - Demo 已实现
...
  /system/user/profile/updatePwd:
    put:
      operationId: changeOwnPassword
      summary: 修改本人密码并撤销旧凭证
      description: 修改本人密码并撤销旧凭证。沿用若依 code/msg 响应；业务错误可能使用 HTTP 200。
      security:
      - bearerAuth: []
      parameters: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
        '400':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
        '401':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
        '403':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
        '404':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
        '409':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
        '413':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
        '429':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
        '503':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
      x-implementation: implemented
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OwnPassword'
      tags:
      - Demo 已实现
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    FrameworkResponse:
      type: object
      additionalProperties: true
      properties:
        code:
          type: integer
        msg:
          type: string
...
    OwnPassword:
      type: object
      additionalProperties: false
      properties:
        oldPassword:
          type: string
        newPassword:
          type: string
          minLength: 12
          maxLength: 64
      required:
      - oldPassword
      - newPassword
    ProfileUpdate:
      type: object
      additionalProperties: false
      properties:
        nickName:
          type: string
          maxLength: 30
        email:
          type: string
          maxLength: 50
        phonenumber:
          type: string
          maxLength: 11
        sex:
          type: string
          enum:
          - '0'
          - '1'
          - '2'
...